The model is changed in a substantial way (everyone please calm down). The reason is that the handling of the pair development/publish wall was not satisfactory. Especially the doubling of most of the fields of a wall and its dev partner is errorprone: If we change specific properties on one wall, (e.g., is_active), the other (e.g., dev) is not updated. While this can be hacked by some workaround, I think the doubling of the fields is a major design flaw.
The new model design proposed in this commit is the following:
The Wall object now hosts all informations on a wall that we do not want
to have changed by users. This is, e.g., the image, the location, the
name (?), the spot.

All informations that change during the workflow we have in mind are
hosted by a WallView object. The WallViews are related many-to-one to a
Wall. The current design is such that there should be exactly two
WallViews, a development view and a publishable view. The two wallviews
are created upon save of the wall and can
be accessed from the wall via properties (dev_view and pub_view). Those
assume of course that the views are created and that there is only one
valid view of each kind.

The WallView hosts at the moment only the is_dev flag and a TextField
that discusses the approach to the wall (Which might be updated by
users).
In principle we can easily move fields from Wall to WallView, this
should be discussed.

The insertfield.sh is updated to work with the new model.
However, the views and templates that are related to the wall model are
BROKEN with this commit. I'll fix that I promise. I thought it is a good idea
to commit these changes asap to avoid any additional "superfluous" work
on the old model.