I like that a lot -- props to Al.
It was however confusing that it did not show up immediately if one
adds routes on the dev walls....
Added them with an prepended asterisk .. open for suggestions though.

And I removed the if stuff -- can the join really break?
Please come back at me for not consulting with you ;P